sys/newlib_syscalls: split filesystem syscalls.
The implementation of read/write/close, etc using VFS is moved to
it's own submodule, newlib_syscalls_riot_vfs.

The stubs for those functions (with the exception of read and write)
now live in newlib_stubs_fs.

The implementaion of read and write that maps every file to the UART
is in newlib_syscalls_riot_serial_rw.
