+++ sys/newlib_stubs/fs.c
@@ -1,0 +1,91 @@
+/*
+ * Copyright (C) 2018 Freie Universität Berlin
+ *               2015 Kaspar Schleiser <kaspar@schleiser.de>
+ *               2014 Freie Universität Berlin
+ *
+ * This file is subject to the terms and conditions of the GNU Lesser General
+ * Public License v2.1. See the file LICENSE in the top level directory for more
+ * details.
+ */
+
+/**
+ * @ingroup    sys_newlib_stubs
+ * @{
+ *
+ * @file
+ * @brief       Stubs for filesystem routines.
+ *
+ * @author      Michael Baar <michael.baar@fu-berlin.de>
+ * @author      Stefan Pfeiffer <pfeiffer@inf.fu-berlin.de>
+ * @author      Hauke Petersen <hauke.petersen@fu-berlin.de>
+ * @author      Kaspar Schleiser <kaspar@schleiser.de>
+ * @author      Juan I Carrano <j.carrano@fu-berlin.de>
+ *
+ * @}
+ */
+
+#include <reent.h>
+#include <errno.h>
+
+/* Stubs to avoid linking errors, these functions do not have any effect */
+int _open_r(struct _reent *r, const char *name, int flags, int mode)
+{
+    (void) name;
+    (void) flags;
+    (void) mode;
+    r->_errno = ENODEV;
+    return -1;
+}
+
+int _close_r(struct _reent *r, int fd)
+{
+    (void) fd;
+    r->_errno = ENODEV;
+    return -1;
+}
+
+_off_t _lseek_r(struct _reent *r, int fd, _off_t pos, int dir)
+{
+    (void) fd;
+    (void) pos;
+    (void) dir;
+    r->_errno = ENODEV;
+    return -1;
+}
+
+int _fstat_r(struct _reent *r, int fd, struct stat *st)
+{
+    (void) fd;
+    (void) st;
+    r->_errno = ENODEV;
+    return -1;
+}
+
+int _stat_r(struct _reent *r, const char *name, struct stat *st)
+{
+    (void) name;
+    (void) st;
+    r->_errno = ENODEV;
+    return -1;
+}
+
+int _unlink_r(struct _reent *r, const char *path)
+{
+    (void) path;
+    r->_errno = ENODEV;
+    return -1;
+}
+
+/*
+ * Riot does not support hard links (that's why the error is ENOSYS instead
+ * of ENODEV.
+ */
+int _link_r(struct _reent *ptr, const char *old_name, const char *new_name)
+{
+    (void)old_name;
+    (void)new_name;
+
+    ptr->_errno = ENOSYS;
+
+    return -1;
+}
