+++ sys/newlib_syscalls_riot/doc.txt
@@ -25,4 +25,6 @@
  *
  * - sys_newlib_syscalls_riot_clock
  * - sys_newlib_syscalls_riot_time
+ * - sys_newlib_syscalls_riot_vfs
+ * - sys_newlib_syscalls_riot_serial_rw
  */
